analysisService/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
get==2019.4.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
MarkupSafe==1.1.1
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.1
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-win_amd64.whl
torchvision==0.4.1
urllib3==1.25.6
Werkzeug==0.16.0
xlrd==1.2.0
